+++ HISTORY.md
@@ -1,3 +1,145 @@
+1.0.3 / 2020-09-29
+===================
+
+  * Reviewed rules:
+    * no-await-in-loop
+    * !! no-console
+    * no-constant-condition
+    * ++ no-dupe-else-if
+    * no-empty
+    * no-extra-parens
+    * ++ no-import-assign
+    * ++ no-loss-of-precision
+    * no-misleading-character-class
+    * ++ no-promise-executor-return
+    * no-prototype-builtins
+    * ++ no-setter-return
+    * no-unreachable
+    * ++ no-unreachable-loop
+    * ++ no-useless-backreference
+    * accessor-pairs
+    * !! complexity
+    * curly
+    * ++ default-case-last
+    * ++ default-param-last
+    * dot-location
+    * dot-notation
+    * eqeqeq
+    * ++ grouped-accessor-pairs
+    * ++ no-constructor-return
+    * ++ no-invalid-this
+    * !! no-magic-numbers
+    * no-multi-spaces
+    * no-new
+    * !! no-param-reassign
+    * no-return-assign
+    * no-unused-expressions
+    * !! no-useless-return
+    * !! no-void
+    * no-warning-comments
+    * ++ prefer-named-capture-group
+    * ++ prefer-regex-literals
+    * wrap-iife
+    * yoda
+    * strict
+    * no-undef
+    * no-unused-vars
+    * no-use-before-define
+    * array-bracket-newline
+    * array-bracket-spacing
+    * array-element-newline
+    * block-spacing
+    * brace-style
+    * ++ camelcase
+    * comma-dangle
+    * comma-style
+    * computed-property-spacing
+    * ++ consistent-this
+    * eol-last
+    * func-call-spacing
+    * func-name-matching
+    * !! func-names
+    * func-style
+    * ++ function-call-argument-newline
+    * !! function-paren-newline
+    * ++ id-denylist
+    * !! id-length
+    * indent
+    * key-spacing
+    * keyword-spacing
+    * linebreak-style
+    * lines-between-class-members
+    * !! max-depth
+    * !! max-len
+    * !! max-lines
+    * multiline-comment-style
+    * multiline-ternary
+    * new-cap
+    * !! newline-per-chained-call
+    * !! no-inline-comments
+    * no-mixed-operators
+    * !! no-multi-assign
+    * no-multiple-empty-lines
+    * !! no-plusplus
+    * !! no-restricted-syntax
+    * no-tabs
+    * no-trailing-spaces
+    * no-underscore-dangle
+    * !! object-curly-newline
+    * object-curly-spacing
+    * object-property-newline
+    * one-var
+    * operator-assignment
+    * operator-linebreak
+    * padded-blocks
+    * !! padding-line-between-statements
+    * ++ prefer-exponentiation-operator
+    * quotes
+    * semi
+    * !! sort-keys
+    * space-before-function-paren
+    * space-in-parens
+    * space-infix-ops
+    * spaced-comment
+    * template-tag-spacing
+    * unicode-bom
+    * arrow-body-style
+    * arrow-parens
+    * ++ no-restricted-exports
+    * !! no-restricted-imports
+    * object-shorthand
+    * prefer-const
+    * prefer-destructuring
+    * !! prefer-rest-params
+    * prefer-template
+    * rest-spread-spacing
+    * !! sort-imports
+    * template-curly-spacing
+    * yield-star-spacing
+    * !! id-blacklist
+    * ++ id-denylist
+    * indent
+    * !! padding-line-between-statements
+    * !! padding-line-between-statements
+    * !! padding-line-between-statements
+    * !! no-restricted-modules
+    * func-call-spacing
+    * strict
+    * no-constant-condition
+    * comma-dangle
+    * strict
+    * no-extra-parens
+    * space-before-function-paren
+    * keyword-spacing
+    * ++ space-before-function-parentheses
+    * space-before-function-paren
+    * keyword-spacing
+    * object-curly-spacing
+    * array-bracket-spacing
+    * keyword-spacing
+    * spaced-comment
+
+  
 1.0.1 - 1.0.2 / 2020-05-21
 ===================